Changed balance value to be int to match with user_dashboard

+++ src/java/Model/User.java
@@ -18,7 +18,7 @@ public class User {
     private LocalDateTime expiresAt;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private double balance;
+    private int balance;
 
     public User() {
     }
@@ -119,11 +119,11 @@ public class User {
         this.updatedAt = updatedAt;
     }
 
-    public double getBalance() {
+    public int getBalance() {
         return balance;
     }
 
-    public void setBalance(double balance) {
+    public void setBalance(int balance) {
         this.balance = balance;
     }
 }
